requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
pydub
python-dotenv
SpeechRecognition
django-imagekit
django-filter
django-tables2
python-ffmpeg